+++ sdk/packages/eternum/src/constants/troops.ts
@@ -1,0 +1,7 @@
+import { ResourcesIds } from "./resources";
+
+export const TROOPS_STAMINAS = {
+  [ResourcesIds.Paladin]: 100,
+  [ResourcesIds.Knight]: 80,
+  [ResourcesIds.Crossbowmen]: 80,
+};
