+++ sdk/packages/eternum/src/utils/index.ts
@@ -1,225 +1,87 @@
-import { CairoCustomEnum } from "starknet";
-import { ResourcesIds } from "../constants";
+import {
+  BUILDING_COSTS,
+  EXPLORATION_COSTS,
+  EternumGlobalConfig,
+  HYPERSTRUCTURE_CONSTRUCTION_COSTS,
+  QUEST_RESOURCES,
+  RESOURCE_BUILDING_COSTS,
+  RESOURCE_INPUTS,
+  RESOURCE_OUTPUTS,
+  ResourceInputs,
+  ResourceOutputs,
+  STRUCTURE_COSTS,
+  HYPERSTRUCTURE_CREATION_COSTS,
+  HYPERSTRUCTURE_TOTAL_COSTS,
+} from "../constants";
+import { Resource } from "../types";
 
-export enum StructureType {
-  Realm = 1,
-  Hyperstructure = 2,
-  Bank = 3,
-  ShardsMine = 4,
-  Settlement = 5,
-}
+export const scaleResourceOutputs = (resourceOutputs: ResourceOutputs, multiplier: number) => {
+  let multipliedCosts: ResourceOutputs = {};
 
-export enum BuildingType {
-  Castle = 1,
-  Resource = 2,
-  Farm = 3,
-  FishingVillage = 4,
-  Barracks = 5,
-  Market = 6,
-  ArcheryRange = 7,
-  Stable = 8,
-  DonkeyFarm = 9,
-  TradingPost = 10,
-  WorkersHut = 11,
-  WatchTower = 12,
-  Walls = 13,
-  Storehouse = 14,
-  Bank = 15,
-  ShardsMine = 16,
-}
-export const MAX_BUILDING_TYPE = 14;
-
-export enum ResourceBuildingType {
-  Wood = 1,
-  Stone = 2,
-  Coal = 3,
-  Copper = 4,
-  Obsidian = 5,
-  Silver = 6,
-  Ironwood = 7,
-  ColdIron = 8,
-  Gold = 9,
-  Hartwood = 10,
-  Diamonds = 11,
-  Sapphire = 12,
-  Ruby = 13,
-  DeepCrystal = 14,
-  Ignium = 15,
-  EtherealSilica = 16,
-  TrueIce = 17,
-  TwilightQuartz = 18,
-  AlchemicalSilver = 19,
-  Adamantine = 20,
-  Mithral = 21,
-  Dragonhide = 22,
-}
-
-export const CombinedBuildingTypes = {
-  ...BuildingType,
-  ...ResourceBuildingType,
+  for (let buildingType in resourceOutputs) {
+    multipliedCosts[buildingType] = resourceOutputs[buildingType] * multiplier;
+  }
+  return multipliedCosts;
 };
 
-// export BUILDING_NONE=0
-// export BUILDING_CASTLE=1
-// export BUILDING_RESOURCE=2
-// export BUILDING_FARM=3
-// export BUILDING_FISHING_VILLAGE=4
-// export BUILDING_BARRACKS=5
-// export BUILDING_MARKET=6
-// export BUILDING_ARCHERY_RANGE=7
-// export BUILDING_STABLE=8
-// export BUILDING_DONKEY_FARM=9
-// export BUILDING_TRADING_POST=10
-// export BUILDING_WORKERS_HUT=11
-// export BUILDING_WATCH_TOWER=12
-// export BUILDING_WALLS=13
-// export BUILDING_STOREHOUSE=14
+export const scaleResourceInputs = (resourceInputs: ResourceInputs, multiplier: number) => {
+  let multipliedCosts: ResourceInputs = {};
 
-export const BuildingEnumToString: { [index: number]: string } = {
-  0: "None",
-  1: "Castle",
-  2: "Resource",
-  3: "Farm",
-  4: "Fishing Village",
-  5: "Barracks",
-  6: "Market",
-  7: "Archery Range",
-  8: "Stable",
-  9: "Donkey Farm",
-  10: "Trading Post",
-  11: "Workers Hut",
-  12: "Watch Tower",
-  13: "Walls",
-  14: "Storehouse",
-  15: "Bank",
-  16: "Shards Mine",
+  for (let buildingType in resourceInputs) {
+    multipliedCosts[buildingType] = resourceInputs[buildingType].map((resourceInput) => ({
+      ...resourceInput,
+      amount: resourceInput.amount * multiplier,
+    }));
+  }
+
+  return multipliedCosts;
 };
 
-export const BuildingStringToEnum = {
-  None: 0,
-  Castle: 1,
-  Resource: 2,
-  Farm: 3,
-  FishingVillage: 4,
-  Barracks: 5,
-  Market: 6,
-  ArcheryRange: 7,
-  Stable: 8,
-  DonkeyFarm: 9,
-  TradingPost: 10,
-  WorkersHut: 11,
-  WatchTower: 12,
-  Walls: 13,
-  Storehouse: 14,
-  Bank: 15,
-  ShardsMine: 16,
+export const scaleResources = (resources: any[], multiplier: number): any[] => {
+  return resources.map((resource) => ({
+    ...resource,
+    amount: resource.amount * multiplier,
+  }));
 };
 
-export function getBuildingType(name: BuildingType): CairoCustomEnum {
-  switch (name) {
-    case BuildingType.Castle:
-      return new CairoCustomEnum({ Castle: {} });
-    case BuildingType.Resource:
-      return new CairoCustomEnum({ Resource: {} });
-    case BuildingType.Farm:
-      return new CairoCustomEnum({ Farm: {} });
-    case BuildingType.FishingVillage:
-      return new CairoCustomEnum({ FishingVillage: {} });
-    case BuildingType.Barracks:
-      return new CairoCustomEnum({ Barracks: {} });
-    case BuildingType.Market:
-      return new CairoCustomEnum({ Market: {} });
-    case BuildingType.ArcheryRange:
-      return new CairoCustomEnum({ ArcheryRange: {} });
-    case BuildingType.Stable:
-      return new CairoCustomEnum({ Stable: {} });
-    case BuildingType.DonkeyFarm:
-      return new CairoCustomEnum({ DonkeyFarm: {} });
-    case BuildingType.TradingPost:
-      return new CairoCustomEnum({ TradingPost: {} });
-    case BuildingType.WorkersHut:
-      return new CairoCustomEnum({ WorkersHut: {} });
-    case BuildingType.WatchTower:
-      return new CairoCustomEnum({ WatchTower: {} });
-    case BuildingType.Walls:
-      return new CairoCustomEnum({ Walls: {} });
-    case BuildingType.Storehouse:
-      return new CairoCustomEnum({ Storehouse: {} });
-    case BuildingType.Bank:
-      return new CairoCustomEnum({ Bank: {} });
-    case BuildingType.ShardsMine:
-      return new CairoCustomEnum({ ShardsMine: {} });
-  }
-}
+export const RESOURCE_BUILDING_COSTS_SCALED: ResourceInputs = scaleResourceInputs(
+  RESOURCE_BUILDING_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
 
-export function getProducedResource(name: BuildingType): number {
-  switch (name) {
-    case BuildingType.Castle:
-      return 0;
-    case BuildingType.Resource:
-      return 0;
-    case BuildingType.Farm:
-      return ResourcesIds.Wheat;
-    case BuildingType.FishingVillage:
-      return ResourcesIds.Fish;
-    case BuildingType.Barracks:
-      return ResourcesIds.Knight;
-    case BuildingType.Market:
-      return 0;
-    case BuildingType.ArcheryRange:
-      return ResourcesIds.Crossbowmen;
-    case BuildingType.Stable:
-      return ResourcesIds.Paladin;
-    case BuildingType.DonkeyFarm:
-      return ResourcesIds.Donkey;
-    case BuildingType.TradingPost:
-      return 0;
-    case BuildingType.WorkersHut:
-      return 0;
-    case BuildingType.WatchTower:
-      return 0;
-    case BuildingType.Walls:
-      return 0;
-    case BuildingType.Storehouse:
-      return 0;
-    case BuildingType.Bank:
-      return 0;
-    case BuildingType.ShardsMine:
-      return 0;
-  }
-}
-
-export enum EntityState {
-  Traveling,
-  WaitingForDeparture,
-  Idle,
-  WaitingToOffload,
-  NotApplicable, // When the entity should not be rendered
-}
-
-export function determineEntityState(
-  nextBlockTimestamp: number | undefined,
-  blocked: boolean | undefined,
-  arrivalTime: number | undefined,
-  hasResources: boolean,
-): EntityState {
-  const isTraveling =
-    !blocked && nextBlockTimestamp !== undefined && arrivalTime !== undefined && arrivalTime > nextBlockTimestamp;
-  const isWaitingForDeparture = blocked;
-  const isIdle = !isTraveling && !isWaitingForDeparture && !hasResources;
-  const isWaitingToOffload = !blocked && !isTraveling && hasResources;
-
-  if (isTraveling) {
-    return EntityState.Traveling;
-  }
-  if (isWaitingForDeparture) {
-    return EntityState.WaitingForDeparture;
-  }
-  if (isIdle) {
-    return EntityState.Idle;
-  }
-  if (isWaitingToOffload) {
-    return EntityState.WaitingToOffload;
-  }
-  return EntityState.Idle; // Default state
-}
+export const RESOURCE_OUTPUTS_SCALED: ResourceOutputs = scaleResourceOutputs(
+  RESOURCE_OUTPUTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const BUILDING_COSTS_SCALED: ResourceInputs = scaleResourceInputs(
+  BUILDING_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const RESOURCE_INPUTS_SCALED: ResourceInputs = scaleResourceInputs(
+  RESOURCE_INPUTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const QUEST_RESOURCES_SCALED: ResourceInputs = scaleResourceInputs(
+  QUEST_RESOURCES,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const EXPLORATION_COSTS_SCALED: Resource[] = scaleResources(
+  EXPLORATION_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const STRUCTURE_COSTS_SCALED: ResourceInputs = scaleResourceInputs(
+  STRUCTURE_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const HYPERSTRUCTURE_CONSTRUCTION_COSTS_SCALED: { resource: number; amount: number }[] = scaleResources(
+  HYPERSTRUCTURE_CONSTRUCTION_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const HYPERSTRUCTURE_CREATION_COSTS_SCALED: { resource: number; amount: number }[] = scaleResources(
+  HYPERSTRUCTURE_CREATION_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
+export const HYPERSTRUCTURE_TOTAL_COSTS_SCALED: { resource: number; amount: number }[] = scaleResources(
+  HYPERSTRUCTURE_TOTAL_COSTS,
+  EternumGlobalConfig.resources.resourceMultiplier,
+);
